+++ nes/apu/apu.go
@@ -1,39 +1,5 @@
 package apu
 
-// https://www.nesdev.org/wiki/APU_Pulse
-var dutyTable = [][]byte{
-	{0, 0, 0, 0, 0, 0, 0, 1},
-	{0, 0, 0, 0, 0, 0, 1, 1},
-	{0, 0, 0, 0, 1, 1, 1, 1},
-	{1, 1, 1, 1, 1, 1, 0, 0},
-}
-
-// https://www.nesdev.org/wiki/APU_Triangle
-var triangleTable = []byte{
-	15, 14, 13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1, 0,
-	0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15,
-}
-
-// https://www.nesdev.org/wiki/APU_Noise
-// NTSC
-var noisePeriodTable = []uint16{
-	4, 8, 16, 32, 64, 96, 128, 160, 202, 254, 380, 508, 762, 1016, 2034, 4068,
-}
-
-// https://www.nesdev.org/wiki/APU_Length_Counter
-var lengthTable = []byte{
-	10, 254, 20, 2, 40, 4, 80, 6, 160, 8, 60, 10, 14, 12, 26, 14,
-	12, 16, 24, 18, 48, 20, 96, 22, 192, 24, 72, 26, 16, 28, 32, 30,
-}
-
-// https://www.nesdev.org/wiki/APU_DMC
-// > These periods are all even numbers because there are 2 CPU cycles in an APU cycle.
-// > A rate of 428 means the output level changes every 214 APU cycles.
-// NTSC
-var dmcPeriodTable = []uint16{
-	428, 380, 340, 320, 286, 254, 226, 214, 190, 160, 142, 128, 106, 84, 72, 54,
-}
-
 // https://www.nesdev.org/wiki/APU_Mixer#Lookup_Table
 var pulseTable [31]float32
 var tndTable [203]float32
@@ -187,7 +153,7 @@ func (apu *APU) WriteNoiseController(val byte) {
 // M---.PPPP	Mode and period (write)
 func (apu *APU) WriteNoiseLoopAndPeriod(val byte) {
 	apu.noise.mode = (val & 0x80) == 0x80
-	apu.noise.timer.period = noisePeriodTable[val&0x0F]
+	apu.noise.loadPeriod(val & 0x0F)
 }
 
 // $400F
@@ -271,294 +237,6 @@ func (apu *APU) output() float32 {
 	pout := pulseTable[apu.pulse1.output()+apu.pulse2.output()]
 	tout := tndTable[3*apu.tnd.output()+2*apu.noise.output()+apu.dmc.output()]
 	return pout + tout
-}
-
-type pulse struct {
-	channel byte
-	lc      lengthCounter
-	el      envelope
-
-	// https://www.nesdev.org/wiki/APU_Sweep
-	sweepEnabled    bool
-	sweepNegate     bool
-	sweepShiftCount byte
-	// > Each sweep unit contains the following: divider, reload flag.
-	sweepDivider divider
-	sweepReload  bool
-
-	duty    byte
-	dutyPos byte
-	timer   timer
-}
-
-func newPulse(channel byte) *pulse {
-	return &pulse{
-		channel: channel,
-		timer:   newTimer(2),
-	}
-}
-
-func (p *pulse) output() byte {
-	if p.lc.value == 0 {
-		return 0
-	}
-	if dutyTable[p.duty][p.dutyPos] == 0 {
-		return 0
-	}
-	if p.timer.period < 8 {
-		return 0
-	}
-	if p.isMuteSweep() {
-		return 0
-	}
-	return p.el.output()
-}
-
-func (p *pulse) targetPeriod() uint16 {
-	// https://www.nesdev.org/wiki/APU_Sweep
-	// > 1. A barrel shifter shifts the channel's 11-bit raw timer period right by the shift count, producing the change amount.
-	// wiki的にはchange amountって言ってるから差分だと思ったんだけど、他実装エミュ見てると、ただのshift結果のコードになってる...
-	delta := p.timer.period >> uint16(p.sweepShiftCount)
-
-	// > 2. If the negate flag is true, the change amount is made negative.
-	// > 3. The target period is the sum of the current period and the change amount.
-	if p.sweepNegate {
-		// > The two pulse channels have their adders' carry inputs wired differently, which produces different results when each channel's change amount is made negative:
-		// > Pulse 1 adds the ones' complement (−c − 1). Making 20 negative produces a change amount of −21.
-		// > Pulse 2 adds the two's complement (−c). Making 20 negative produces a change amount of −20.
-		if p.channel == 1 {
-			return p.timer.period - delta - 1
-		} else {
-			return p.timer.period - delta
-		}
-	} else {
-		return p.timer.period + delta
-	}
-}
-
-func (p *pulse) isMuteSweep() bool {
-	// > Two conditions cause the sweep unit to mute the channel:
-	// > 1. If the current period is less than 8, the sweep unit mutes the channel.
-	// > 2. If at any time the target period is greater than $7FF, the sweep unit mutes the channel.
-	return p.timer.period < 8 || p.targetPeriod() > 0x7FF
-}
-
-func (p *pulse) tickSweep() {
-	// > 1. If the divider's counter is zero, the sweep is enabled, and the sweep unit is not muting the channel: The pulse's period is adjusted.
-	// > 2. If the divider's counter is zero or the reload flag is true: The counter is set to P and the reload flag is cleared. Otherwise, the counter is decremented.
-
-	if p.sweepDivider.tick() && p.sweepEnabled && !p.isMuteSweep() {
-		p.timer.period = p.targetPeriod()
-	}
-
-	if p.sweepReload {
-		p.sweepReload = false
-		p.sweepDivider.reload()
-	}
-}
-
-func (p *pulse) tickTimer() {
-	if p.timer.tick() {
-		p.dutyPos = (p.dutyPos - 1) & 7
-	}
-}
-
-func (p *pulse) tickEnvelope() {
-	p.el.tick()
-}
-
-func (p *pulse) tickLengthCounter() {
-	p.lc.tick()
-}
-
-type triangle struct {
-	seqPos              byte
-	lc                  lengthCounter
-	linearCounterCtrl   bool
-	linearCounter       byte
-	linearCounterPeriod byte
-	linearCounterReload bool
-
-	timer timer
-}
-
-func newTriangle() *triangle {
-	return &triangle{
-		timer: newTimer(1),
-	}
-}
-
-func (t *triangle) output() byte {
-	if t.lc.value == 0 {
-		return 0
-	}
-	// todo
-	return 0
-}
-
-func (t *triangle) tickTimer() {
-	if t.timer.tick() {
-		if t.lc.value > 0 && t.linearCounter > 0 {
-			t.seqPos = (t.seqPos + 1) % 32
-		}
-	}
-}
-
-func (t *triangle) tickLinearCounter() {
-	// > 1. If the linear counter reload flag is set, the linear counter is reloaded with the counter reload value, otherwise if the linear counter is non-zero, it is decremented.
-	// > 2. If the control flag is clear, the linear counter reload flag is cleared.
-	if t.linearCounterReload {
-		t.linearCounter = t.linearCounterPeriod
-	} else if t.linearCounter > 0 {
-		t.linearCounter--
-	}
-	if !t.linearCounterCtrl {
-		t.linearCounterReload = false
-	}
-}
-
-func (t *triangle) tickLengthCounter() {
-	t.lc.tick()
-}
-
-type noise struct {
-	lc            lengthCounter
-	el            envelope
-	mode          bool
-	shiftRegister uint16
-	timer         timer
-}
-
-func newNoise() *noise {
-	return &noise{
-		timer: newTimer(2),
-	}
-}
-
-func (n *noise) output() byte {
-	if n.lc.value == 0 {
-		return 0
-	}
-	if (n.shiftRegister & 1) == 1 {
-		return 0
-	}
-	return n.el.output()
-}
-
-func (n *noise) tickTimer() {
-	if n.timer.tick() {
-		fb := uint16(0)
-		if n.mode {
-			fb = (n.shiftRegister & 0x01) ^ ((n.shiftRegister >> 6) & 0x01)
-		} else {
-			fb = (n.shiftRegister & 0x01) ^ ((n.shiftRegister >> 1) & 0x01)
-		}
-		n.shiftRegister >>= 1
-		n.shiftRegister |= (fb << 14)
-	}
-}
-
-func (n *noise) tickEnvelope() {
-	n.el.tick()
-}
-
-func (n *noise) tickLengthCounter() {
-	n.lc.tick()
-}
-
-type dmc struct {
-	enabled      bool
-	irqEnabled   bool
-	loop         bool
-	freq         byte
-	counter      byte
-	sampleAddr   byte
-	sampleLength byte
-}
-
-func newDMC() *dmc {
-	return &dmc{}
-}
-
-func (d *dmc) setEnabled(v bool) {
-	if !v {
-		// todo
-	}
-	d.enabled = v
-}
-
-func (d *dmc) output() byte {
-	// todo
-	return 0
-}
-
-type envelope struct {
-	constantVolume    bool
-	loop              bool
-	start             bool
-	divider           divider
-	decayLevelCounter byte
-}
-
-func (e *envelope) output() byte {
-	// https://www.nesdev.org/wiki/APU_Envelope
-	// > The envelope unit's volume output depends on the constant volume flag:
-	// > if set, the envelope parameter directly sets the volume, otherwise the decay level is the current volume.
-	if e.constantVolume {
-		// > bits 3-0	---- VVVV	Used as the volume in constant volume (C set) mode.
-		// > Also used as the reload value for the envelope's divider (the period becomes V + 1 quarter frames).
-		return byte(e.divider.period)
-	} else {
-		return e.decayLevelCounter
-	}
-}
-
-func (e *envelope) tick() {
-	// https://www.nesdev.org/wiki/APU_Envelope
-	// > When clocked by the frame counter, one of two actions occurs:
-	// > if the start flag is clear, the divider is clocked,
-	// > otherwise the start flag is cleared, the decay level counter is loaded with 15, and the divider's period is immediately reloaded.
-	if e.start {
-		e.start = false
-		e.decayLevelCounter = 15
-		e.divider.reload()
-	} else if e.divider.tick() {
-		// > When the divider is clocked while at 0, it is loaded with V and clocks the decay level counter.
-		// > Then one of two actions occurs: If the counter is non-zero, it is decremented, otherwise if the loop flag is set, the decay level counter is loaded with 15.
-		if e.decayLevelCounter > 0 {
-			e.decayLevelCounter--
-		} else if e.loop {
-			e.decayLevelCounter = 15
-		}
-	}
-}
-
-type lengthCounter struct {
-	enabled bool
-	halt    bool
-	value   byte
-}
-
-func (lc *lengthCounter) setEnabled(v bool) {
-	if v {
-		lc.enabled = true
-	} else {
-		lc.enabled = false
-		lc.value = 0
-	}
-}
-
-func (lc *lengthCounter) load(v byte) {
-	// > If the enabled flag is set, the length counter is loaded with entry L of the length table
-	if lc.enabled {
-		lc.value = lengthTable[v]
-	}
-}
-
-func (lc *lengthCounter) tick() {
-	if lc.value > 0 && !lc.halt {
-		lc.value--
-	}
 }
 
 type timer struct {
